refactor(tests): Split out generic session tests
