+++ tests/test_frames.py
@@ -8,9 +8,7 @@ import pytest  # type: ignore
 
 import nixnet
 from nixnet import _frames
-from nixnet import _utils
 from nixnet import constants
-from nixnet import errors
 from nixnet import types
 
 
@@ -182,130 +180,3 @@ def test_singlepoint_loopback(nixnet_in_interface, nixnet_out_interface):
             assert len(expected_frames) == len(actual_frames)
             for i, (expected, actual) in enumerate(zip(expected_frames, actual_frames)):
                 assert_can_frame(i, expected, actual)
-
-
-@pytest.mark.integration
-def test_session_properties(nixnet_out_interface):
-    """Verify Session properties.
-
-    Ideally, mutable properties would be set to multiple values and we'd test
-    for the intended side-effect.  That'll be a massive undertaking.  For now,
-    ensure they are settable and getttable.
-    """
-    database_name = 'NIXNET_example'
-    cluster_name = 'CAN_Cluster'
-    frame_name = 'CANEventFrame1'
-
-    with nixnet.FrameOutQueuedSession(
-            nixnet_out_interface,
-            database_name,
-            cluster_name,
-            frame_name) as output_session:
-        print(output_session.time_current)
-        assert output_session.state == constants.SessionInfoState.STOPPED
-        print(output_session.can_comm)
-
-        assert output_session.database_name == database_name
-        assert output_session.cluster_name == cluster_name
-        assert output_session.mode == constants.CreateSessionMode.FRAME_OUT_QUEUED
-        assert output_session.application_protocol == constants.AppProtocol.NONE
-        assert output_session.protocol == constants.Protocol.CAN
-
-        assert output_session.auto_start
-        output_session.auto_start = False
-        assert not output_session.auto_start
-
-        print(output_session.num_pend)
-        print(output_session.num_unused)
-        print(output_session.frames.payld_len_max)
-
-        print(output_session.queue_size)
-        output_session.queue_size = 2040
-        assert output_session.queue_size == 2040
-
-
-@pytest.mark.integration
-def test_session_properties_transition(nixnet_out_interface):
-    """Verify Session properties relationship to session start/stop."""
-    database_name = 'NIXNET_example'
-    cluster_name = 'CAN_Cluster'
-    frame_name = 'CANEventFrame1'
-
-    with nixnet.FrameOutQueuedSession(
-            nixnet_out_interface,
-            database_name,
-            cluster_name,
-            frame_name) as output_session:
-        with pytest.raises(errors.XnetError):
-            print(output_session.time_start)
-            print(output_session.time_communicating)
-        assert output_session.state == constants.SessionInfoState.STOPPED
-
-        output_session.start()
-
-        print(output_session.time_start)
-        print(output_session.time_communicating)
-        assert output_session.state == constants.SessionInfoState.STARTED
-
-        output_session.stop()
-
-        with pytest.raises(errors.XnetError):
-            print(output_session.time_start)
-            print(output_session.time_communicating)
-        assert output_session.state == constants.SessionInfoState.STOPPED
-
-
-def test_parse_can_comm_bitfield():
-    """A part of Session.can_comm"""
-    comm = _utils.parse_can_comm_bitfield(0)
-    assert comm == types.CanComm(
-        constants.CanCommState.ERROR_ACTIVE,
-        tcvr_err=False,
-        sleep=False,
-        last_err=constants.CanLastErr.NONE,
-        tx_err_count=0,
-        rx_err_count=0)
-
-    comm = _utils.parse_can_comm_bitfield(0xFFFFF6F3)
-    assert comm == types.CanComm(
-        constants.CanCommState.INIT,
-        tcvr_err=True,
-        sleep=True,
-        last_err=constants.CanLastErr.CRC,
-        tx_err_count=255,
-        rx_err_count=255)
-
-
-@pytest.mark.integration
-def test_frames_container(nixnet_in_interface):
-    database_name = 'NIXNET_example'
-    cluster_name = 'CAN_Cluster'
-    frame_name = 'CANEventFrame1'
-
-    with nixnet.FrameInQueuedSession(
-            nixnet_in_interface,
-            database_name,
-            cluster_name,
-            frame_name) as input_session:
-        assert frame_name in input_session.frames
-        assert 0 in input_session.frames
-
-        assert len(input_session.frames) == 1
-        frames = list(input_session.frames)
-        assert len(frames) == 1
-        frame = frames[0]
-
-        assert str(frame) == frame_name
-        assert int(frame) == 0
-
-        assert frame == input_session.frames[0]
-        assert frame == input_session.frames[frame_name]
-        with pytest.raises(IndexError):
-            input_session.frames[1]
-        with pytest.raises(KeyError):
-            input_session.frames["<random>"]
-
-        assert frame == input_session.frames.get(0)
-        assert frame == input_session.frames.get(frame_name)
-        assert input_session.frames.get(1) is None
-        assert input_session.frames.get("<random>") is None
